src/backend/requirements.txt
asyncpg~=0.29.0
fastapi~=0.108.0
uvicorn~=0.25.0
psycopg2-binary~=2.9.5
SQLAlchemy~=2.0.0
fastapi-pagination~=0.12.14
alembic~=1.13.0
fastapi-sso~=0.15.0
python-socketio~=5.11.4
pytest-asyncio~=0.24.0
aiohttp~=3.10.10
msal~=1.31.0

# dev
pytest~=8.3.1
httpx~=0.26.0
expiringdict
bcrypt
python-dotenv
pydantic-to-typescript~=1.0.1